Solucionando conflictos con los fragments
