--- app/src/main/java/com/example/sky/CerrarSesionFragment.java
@@ -1,35 +1,0 @@
-package com.example.sky;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-public class CerrarSesionFragment extends Fragment {
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        // Inflar el diseño del fragmento
-        View view = inflater.inflate(R.layout.fragment_cerrar_sesion, container, false);
-
-        TextView tvMessage = view.findViewById(R.id.tvMessage);
-        Button btnConfirmar = view.findViewById(R.id.btnConfirmar);
-
-        btnConfirmar.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                // Redirigir a la pantalla de inicio de sesión.
-                requireActivity().finish(); // Cierra la actividad actual
-            }
-        });
-
-        return view;
-    }
-}
