+++ app/src/main/java/com/example/sky/ui/cerrar_sesion/CerrarSesionFragment.java
@@ -1,34 +1,38 @@
 package com.example.sky.ui.cerrar_sesion;
 
+
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
 
 import com.example.sky.R;
 
-
 public class CerrarSesionFragment extends Fragment {
 
-    public CerrarSesionFragment() {
-
-    }
-
+    @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_cerrar_sesion, container, false);
-    }
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        // Inflar el diseño del fragmento
+        View view = inflater.inflate(R.layout.fragment_cerrar_sesion, container, false);
 
+        TextView tvMessage = view.findViewById(R.id.tvMessage);
+        Button btnConfirmar = view.findViewById(R.id.btnConfirmar);
 
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
+        btnConfirmar.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // Redirigir a la pantalla de inicio de sesión.
+                requireActivity().finish(); // Cierra la actividad actual
+            }
+        });
+
+        return view;
     }
 }
